requirements.txt
altgraph==0.17.3
click==8.1.3
colorama==0.4.6
Flask==2.2.2
future==0.18.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
pefile==2022.5.30
playwright==1.29.0
protobuf==3.20.3
pyee==9.0.4
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pywin32-ctypes==0.2.0
typing_extensions==4.4.0
Werkzeug==2.2.2
